Add shift InvoiceRow to Invoice package
