+++ packages/invoices/src/OutboundShipment/ListView/Toolbar.tsx
@@ -1,8 +1,6 @@
 import React, { FC, useEffect, useRef } from 'react';
-
 import {
   useNotification,
-  InvoiceRow,
   DropdownMenu,
   DropdownMenuItem,
   useTranslation,
@@ -11,8 +9,9 @@ import {
   AppBarContentPortal,
   InputWithLabelRow,
   BasicTextInput,
+  FilterController,
 } from '@openmsupply-client/common';
-import { FilterController } from '@openmsupply-client/common/src/hooks/useFilterBy';
+import { InvoiceRow } from '../../types';
 
 export const Toolbar: FC<{
   onDelete: (toDelete: InvoiceRow[]) => void;
